+++ packages/Rfmalloc/src/fmalloc.cpp
@@ -10,5 +10,6 @@
 #include "fmalloc_alp.inc"
 #include "fmalloc_sparse.inc"
 #include "fmalloc_bed.inc"
+#include "fmalloc_dosage.inc"
 #include "fmalloc_insitu.inc"
 #include "fmalloc_api.inc"